requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.33
botocore==1.34.33
Brotli==1.1.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cssselect2==0.7.0
Django==4.2.9
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-environ==0.11.2
django-ses==3.5.2
django-storages==1.14.2
django-timezone-field==6.1.0
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
fonttools==4.47.2
gunicorn==21.2.0
html5lib==1.1
idna==3.6
jmespath==1.0.1
kombu==5.3.5
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.8.0
pyphen==0.14.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
tzdata==2023.4
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
weasyprint==60.2
webencodings==0.5.1
zopfli==0.2.3
